requirements.txt
tensorflow==2.16.1
streamlit==1.35.0
pillow==10.3.0
numpy==1.26.4